Split integration-test code into mods
